refactor(bridge):delete private bridge index field
- 불필요한 다리 인덱스 필드 삭제, userBridge의 사이즈로 읽어오도록 변경
